Fixed bugs in all implementations
Fixed Csharp implementation not handling bools in DynamoDB properly.
They need to be wrapped in the DynamoDBBool object to serialize to bools
instead of numbers.

Fixed Csharp deployment script.

Fixed wrong error code check in go when checking object presence, don't
know why NoSuchKey does not work, but S3 returns NotFound.

Fixed serialization of users in go, which was serialized into list
instead of stringset. Need to specify it directly in the tag, as
stringset is supposed to contain unique values, which AWS SDK cannot
guarantee by itself without us specifying it explicitly.

Fixed objectCheck, which was returning false always, as the object
either existed or it returned with error, not both at once.

Remove debug prints from JAVA implementation, it is really just that
slow. Needed to incease the timeout to 30 seconds ,as java
implementation runs up to 23 seconds on first run.

Fixed java path manipulation, which was incorrectly processing the
absolute path into normalized path.

Fixed Python conditions when deciding what to do based on objectCheck
and entry create. Was missing one condition which sent it into the wrong
branch.

Fixed typescript path manipulation. I tried to optimize it, which just
broke it, so I reverted back to the Anzu implementation.

+++ python/main.py
@@ -31,7 +31,7 @@ def create_file(path: FilePath):
 
     if object_exists and entry_existed:
         raise errors.ConflictError("File already exists")
-    elif not entry_existed:
+    elif object_exists and not entry_existed:
         # Delete the entry again
         try:
             permDB.delete_master_entry(path)
